requirements.txt
altgraph==0.17.2
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.2.7
auto-py-to-exe==2.15.0
beautifulsoup4==4.8.2
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.10
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
Django==3.0.5
django-tinymce4-lite==1.7.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Eel==0.12.4
entrypoints==0.3
fonttools==4.28.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
grpcio==1.16.1
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
jsmin==2.2.2
keyring==18.0.1
kiwisolver==1.3.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.0
nemo-emblems==4.6.0
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
numpy==1.21.4
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
paddle-client==1.0.0
PAM==0.4.2
pandas==1.3.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pypi==2.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-xapp==2.0.1
python-xlib==0.23
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
scipy==1.7.3
SecretStorage==2.3.1
setproctitle==1.1.10
setuptools-scm==6.3.2
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.1
systemd-python==234
tinycss2==1.0.2
tomli==1.2.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
vboxapi==1.0
wadllib==1.3.3
webencodings==0.5.1
whichcraft==0.6.1
xkit==0.0.0
zeroconf==0.24.4
zope.event==4.5.0
zope.interface==5.4.0
